requirements.txt
requests
beautifulsoup4
slack_sdk
python-dotenv
python-telegram-bot==13.15
